pyproject.toml
[project]
name = "tylko-recruitment-enhanced-repo"
version = "0.1.0"
description = "Tylko recruitment enhanced repo"
authors = [
    {name = "chilledsnake",email = "cwroblewski@o2.pl"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "markdown (>=3.8.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "ruff (>=0.12.1,<0.13.0)",
    "isort (>=6.0.1,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
